api/requirements.txt
asgiref==3.5.2
asn1crypto==1.5.1
cbor2==5.4.3
cffi==1.15.1
cryptography==37.0.4
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.10.1
pyOpenSSL==22.0.0
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
webauthn==1.6.0
